% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sunsentiments.R
\name{get_sunsentiments}
\alias{get_sunsentiments}
\title{Get a tidy data frame of a single sentiment lexicon}
\usage{
get_sunsentiments(lexicon = c("sunscReen"))
}
\arguments{
\item{lexicon}{The sentiment lexicon to retrieve;
either "afinn", "bing", "nrc", or "loughran" or "sunscReen"}
}
\description{
Get specific sentiment lexicons in a tidy format, with one row per word,
in a form that can be joined with a one-word-per-row dataset.
The \code{"bing"} option comes from the included \code{\link{sentiments}}
data frame, and others call the relevant function in the \pkg{textdata}
package.
}
