% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sunsentiments.R
\name{lexicon_sunscReen}
\alias{lexicon_sunscReen}
\title{lexicon_sunscReen}
\usage{
lexicon_sunscReen()
}
\value{
A \code{dataframe} of sentiments
}
\description{
lexicon_sunscReen
}
\section{Functions}{
\itemize{
\item \code{lexicon_sunscReen}: is used in getsunsentiments
}}

