% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sunsentiments.R
\name{get_sunsentiments}
\alias{get_sunsentiments}
\title{get_sunsentiments}
\usage{
get_sunsentiments(lexicon = c("sunscReen"))
}
\arguments{
\item{lexicon}{A \code{character} (character) used to name the lexicon}
}
\value{
A \code{dataframe} of words with sentiments
}
\description{
get_sunsentiments
}
\section{Functions}{
\itemize{
\item \code{get_sunsentiments}: gets sentiments about sunscreens depending on words
}}

\examples{
get_sunsentiments("sunscReen")
}
